split ScrapeRoomTask into its own module
