+++ calendar/src/scraping/mod.rs
@@ -1,2 +1,3 @@
 pub mod continous_scraping;
+mod task;
 mod tumonline_calendar;
